% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpCompare-package.R
\docType{package}
\name{fpCompare-package}
\alias{fpCompare}
\alias{fpCompare-package}
\title{fpCompare: Reliable Comparison of Floating Point Numbers}
\description{
Comparisons of floating point numbers are problematic due to errors associated with the binary representation of decimal numbers. Despite being aware of these problems, people still use numerical methods that fail to account for these and other rounding errors (this pitfall is the first to be highlighted in Circle 1 of Burns (2012) 'The R Inferno' \url{https://www.burns-stat.com/pages/Tutor/R_inferno.pdf}). This package provides new relational operators useful for performing floating point number comparisons with a set tolerance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fpcompare.predictiveecology.org}
  \item \url{https://github.com/PredictiveEcology/fpCompare}
  \item Report bugs at \url{https://github.com/PredictiveEcology/fpCompare/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})

Authors:
\itemize{
  \item Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})
}

Other contributors:
\itemize{
  \item His Majesty the King in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]
}

}
\keyword{internal}
